Update the test cases to only check the best match
cpe_search is only deterministic in all of its results if the CPE
dict is in the same order. With the new building mechanism via the
NVD's API, the CPE dict is not always in the same order. Therefore,
the only part of the result that is guaranteed to be deterministic
now is the best match.