Add weighing of terms for cosine similarity
For the cosine similarity comparison between user query and
CPE entries, an exponential decay function is now used for
assigning decreasing weights to terms according to their position
in a CPE entry.

This has the effect that terms that appear later
in a CPE entry contribute less to the comparison. It solves a
general problem, wherein wrong software CPEs were returned if only
one word between user query and CPE entry matched, for example.