Change search approach and alternate queries creation
Before, cpe_search tried to avoid returning too many similar CPEs based
on the first two CPE fields alone (i.e. the product name). This led to
an issue, wherein CPEs with only a product would block CPEs with a
version number from another product because their similarity score was
higher.

Now, cpe_search tries to avoid returning too many similar CPEs based on
the number of non-wildcard fields in a CPE. Also, the number of similar
CPEs that are kept track of during the search is not limited anymore.
This is accomplished by using a divide and conquer approach for the
search based on a CPE's "class".

Furthermore, simplify addition of new correcting alternative queries.